Update tests with mlflow_registry_uri
