+++ tests/conftest.py
@@ -338,8 +338,13 @@ def move_artifact_storage_dir():
 
 
 @pytest.fixture
-@patch("lineapy.api.api._try_write_to_pickle", return_value=None)
-@patch("lineapy.api.api._pickle_name", return_value="pickle-sample.pkl")
+@patch(
+    "lineapy.api.artifact_serializer._try_write_to_pickle", return_value=None
+)
+@patch(
+    "lineapy.api.artifact_serializer._pickle_name",
+    return_value="pickle-sample.pkl",
+)
 def housing_tracer(_pickle_name, try_write_to_pickle, execute):
     tests_dir = Path(__file__).parent
